Fix specs and rework database loading
The older version of combustion used by awesome_nested_set doesn't
appear to be compatible with rails 7, and the latest version of
combustion no longer has an API that lets you pass in a config.

So this commit drops the dependency on combustion and just uses the
rails task for database setup and teardown.